avoid size_t comparison warn

+++ include/lwpr.hh
@@ -52,13 +52,13 @@ inline Eigen::VectorXd doubleVecToEigen(const doubleVec& ret) {
 
 inline Eigen::MatrixXd doubleVecToEigen(const std::vector<doubleVec>& tmp) {
       Eigen::MatrixXd ret(tmp.size(), tmp[0].size());
-      for(int i=0;i<tmp.size();i++) ret.row(i) = Eigen::Map<Eigen::VectorXd>(const_cast<double*>(tmp[i].data()),tmp[i].size());
+      for(size_t i=0;i<tmp.size();i++) ret.row(i) = Eigen::Map<Eigen::VectorXd>(const_cast<double*>(tmp[i].data()),tmp[i].size());
       return ret;
    }
 
 inline std::vector<Eigen::VectorXd> doubleVecToEigenVec(const std::vector<doubleVec>& tmp) {
 	std::vector<Eigen::VectorXd> ret(tmp.size());
-	for (int i = 0; i<tmp.size(); i++) ret[i] = Eigen::Map<Eigen::VectorXd>(const_cast<double*>(tmp[i].data()), tmp[i].size());
+	for (size_t i = 0; i<tmp.size(); i++) ret[i] = Eigen::Map<Eigen::VectorXd>(const_cast<double*>(tmp[i].data()), tmp[i].size());
 	return ret;
 }
